back/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
albumentations==0.5.0     # via -r requirements.in
certifi==2020.6.20        # via matplotlib, requests
chardet==3.0.4            # via requests
click==7.1.2              # via flask
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via networkx
flask==1.1.2              # via -r requirements.in
future==0.18.2            # via torch
idna==2.10                # via requests
imageio==2.9.0            # via -r requirements.in, imgaug, scikit-image
imgaug==0.4.0             # via albumentations
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
kiwisolver==1.2.0         # via matplotlib
markupsafe==1.1.1         # via jinja2
matplotlib==3.3.2         # via imgaug, scikit-image
munch==2.5.0              # via pretrainedmodels
networkx==2.5             # via scikit-image
numpy==1.19.2             # via -r requirements.in, albumentations, imageio, imgaug, matplotlib, opencv-python, opencv-python-headless, pywavelets, scikit-image, scipy, tifffile, torch, torchvision
opencv-python-headless==4.4.0.44  # via albumentations
opencv-python==4.4.0.44   # via imgaug
pillow==8.0.1             # via -r requirements.in, imageio, imgaug, matplotlib, scikit-image, torchvision
pretrainedmodels==0.7.4   # via -r requirements.in
pyparsing==2.4.7          # via matplotlib
python-dateutil==2.8.1    # via matplotlib
pywavelets==1.1.1         # via scikit-image
pyyaml==5.3.1             # via albumentations
requests==2.24.0          # via -r requirements.in
scikit-image==0.17.2      # via albumentations, imgaug
scipy==1.5.3              # via albumentations, imgaug, scikit-image
shapely==1.7.1            # via imgaug
six==1.15.0               # via cycler, imgaug, munch, python-dateutil
tifffile==2020.9.3        # via scikit-image
torch==1.6.0              # via -r requirements.in, pretrainedmodels, torchvision
torchvision==0.7.0        # via -r requirements.in, pretrainedmodels
tqdm==4.50.2              # via pretrainedmodels
urllib3==1.25.11          # via requests
waitress==1.4.4           # via -r requirements.in
werkzeug==1.0.1           # via flask
